+++ alinkjs/packages/assetlink-default-plugins/plugins/DefaultPrecacher.alink.js
@@ -51,7 +51,9 @@ export default class DefaultPrecacher {
           .sort('-changed')
           .page({ offset: 0, limit: 100 })
           .options({ include });
-      }));
+      }), {
+        priority: 10,
+      });
 
       // TODO: Consider also precaching recently changed and recent/upcoming logs