+++ alinkjs/packages/assetlink/.babelrc.js
@@ -1,6 +1,7 @@
 module.exports = {
   presets: [],
   plugins: [
+    ["@babel/plugin-syntax-decorators", { "decoratorsBeforeExport": true }],
     [
       require.resolve('babel-plugin-module-resolver'),
       {