+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Perform precaching with lower priority than other JSON:API requests
+
 ## [0.7.5] - 2022-10-18
 
 ### Changed